requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
cssbeautifier==1.14.9
Django==4.2.5
djlint==1.34.0
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.14.9
json5==0.9.14
pathspec==0.11.2
Pillow==10.0.1
PyYAML==6.0.1
regex==2023.8.8
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
